[refactor](typo): use-action should be course-action. file name changed
